+++ src/includes/traits/Facades/TransientShortlink.php
@@ -1,0 +1,26 @@
+<?php
+declare (strict_types = 1);
+namespace WebSharks\WpSharks\Core\Traits\Facades;
+
+use WebSharks\WpSharks\Core\Classes;
+use WebSharks\WpSharks\Core\Interfaces;
+use WebSharks\WpSharks\Core\Traits;
+#
+use WebSharks\Core\WpSharksCore\Classes as CoreClasses;
+use WebSharks\Core\WpSharksCore\Classes\Core\Base\Exception;
+use WebSharks\Core\WpSharksCore\Interfaces as CoreInterfaces;
+use WebSharks\Core\WpSharksCore\Traits as CoreTraits;
+#
+use function assert as debug;
+use function get_defined_vars as vars;
+
+trait TransientShortlink
+{
+    /**
+     * @since 160704 Transient shortlinks.
+     */
+    public static function transientShortlink(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->§TransientShortlink->__invoke(...$args);
+    }
+}